Решение задачи с массивами и выделением памяти long long  в файле malloc

+++ mallocArray/Malloc.c
@@ -3,20 +3,20 @@
 
 int main() {
 
-    int sizeNumber; 
+    long int sizeNumber; 
 
     printf("Enter number: ");
-    scanf_s("%d", &sizeNumber);
+    scanf_s("%ld", &sizeNumber);
 
-    int* arrs = malloc(sizeNumber * sizeof(int));
+    long long int* arrs = malloc(sizeNumber * sizeof(long long int));
 
     for(int i = 0; i < sizeNumber; i++) {
         printf("Enter number array: %d ", i + 1);
-        scanf_s("%d", &arrs[i]);
+        scanf_s("%lld", &arrs[i]);
     }
 
     for(int j = 0; j < sizeNumber; j++) {
-        printf("Array[%d]: %d\n", j + 1, arrs[j]);
+        printf("Array[%d]: %lld\n", j + 1, arrs[j]);
     }
